[FELIX-4369] [FELIX-4370] Initial work towards supporting the OSGi Repository 1.0 spec
This commit provides a start for support of reading the spec-compliant Repository XML. The API for reading the XML hasn't changed, however when an XML document is encountered with the http://www.osgi.org/xmlns/repository/v1.0.0 name space it is parsed as a spec-compliant Repository XML.
This commit provides a start for supporting the Repository Service API, the actual Repository Service is not yet registered, but an implementation of this API can be obtained by wrapping the RepositoryAdmin object with the OSGiRepositoryImpl class.

Part of the work is based on and inspired by the OBR-Repository mapping as used by the Apache Aries Subsystems project (thanks!).

This work is by no means finished, it's just a start. Old clients should still work (at least the existing unit tests all pass) and new unit tests were added.




git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@1563841 13f79535-47bb-0310-9956-ffa450edef68